+++ lib/px/utils/pxconstants.lua
@@ -3,7 +3,7 @@
 ----------------------------------------------
 
 local _M = {
-    MODULE_VERSION = "NGINX Module v7.2.1",
+    MODULE_VERSION = "NGINX Module v7.3.0",
     RISK_PATH = "/api/v3/risk",
     CAPTCHA_PATH = "/api/v2/risk/captcha",
     ACTIVITIES_PATH = "/api/v1/collector/s2s",
@@ -20,7 +20,7 @@ local _M = {
     HSC_BLOCK_ACTION = 'hsc',
     HSC_DRC_PROPERTY = 7190,
     HSC_BLOCK_TYPE = 'pxHypeSaleChallenge',
-    GRAPHQL_PATH = "graphql",
+    GRAPHQL_PATH = "/graphql",
     GRAPHQL_QUERY = "query",
     GRAPHQL_MUTATION = "mutation",
     JSON_CONTENT_TYPE = "application/json",
@@ -34,7 +34,17 @@ local _M = {
     EMAIL_ADDRESS_REGEX = "[A-Za-z0-9%.%%%+%-]+@[A-Za-z0-9%.%%%+%-]+%.%w%w%w?%w?",
     GMAIL_DOMAIN = "gmail.com",
     BACKUP_CAPTCHA_HOST = "https://captcha.px-cloud.net",
-    CTS_COOKIE = "pxcts"
+    CTS_COOKIE = "pxcts",
+    CORS_HEADER_KEY = "Access-Control-Allow-Origin",
+    CORS_ALLOW_CREDENTIALS_HEADER_KEY = "Access-Control-Allow-Credentials",
+    CORS_ACCESS_CONTROL_ALLOW_METHODS_KEY = "Access-Control-Allow-Methods",
+    CORS_ACCESS_CONTROL_REQUEST_METHOD_HEADER = "Access-Control-Request-Method",
+    CORS_ACCESS_CONTROL_ALLOW_METHODS_VALUE = "GET,POST,OPTIONS",
+    CORS_ACCESS_CONTROL_ALLOW_HEADERS_KEY = "Access-Control-Allow-Headers",
+    CORS_ACCESS_CONTROL_ALLOW_HEADERS_VALUE = "Content-Type,Authorization",
+    ORIGIN_HEADER_NAME = "origin",
+    OPTIONS_METHOD = "OPTIONS"
+
 }
 
 return _M